airflow/requirements.txt
apache-airflow==2.2.0
virtualenv
python-dotenv
pymysql
pandas
mysql-connector-python>=8.0.11, <9
mysqlclient>=1.3.6,<3
apache-airflow[mysql]

